Switched output to meter space rather than coordinate space
Changed behavior of application to send geometry in meter space, rather
than WGS84 coordinates. The original lat lng coordinates are preserved
in the attributes property of the geometry, along with the reference
point. No longer need flow block to convert to meter space, though we
will need a block that can align multiple outputs to the same reference
point.